feat: updated icons
Add icons from the Papirus icon theme with a slight modification
increasing the contranst of the icons with their border.

This is for making the icons more adaptable to the opposite theme
and it is necessary since the hicolor themes only accepts one of the
themes variants, currently we are using the dark one.

With this change even the one for dark themes will look good in white themes,
and vice-versa.